feat(fuselage)!: Remove `ContextualbarV2` in favor of `Contextualbar`
Drop the `ContextualbarV2` component family and move its implementation
into the `Contextualbar` components, which are no longer deprecated.
