+++ .changeset/olive-pandas-repeat.md
@@ -1,0 +1,7 @@
+---
+'@rocket.chat/fuselage': minor
+---
+
+feat(fuselage)!: Remove the `ContextualbarV2` component family and replace the `Contextualbar` component family with its implementation
+
+The `ContextualbarV2*` components and types are no longer exported; use the `Contextualbar*` ones instead, which now carry the V2 look and behavior.
