+++ packages/fuselage/src/components/Contextualbar/Contextualbar.spec.tsx
@@ -7,6 +7,8 @@ import * as stories from './Contextualbar.stories';
 
 expect.extend(toHaveNoViolations);
 
+const { WithLongTitle } = composeStories(stories);
+
 const testCases = Object.values(composeStories(stories)).map((Story) => [
   Story.storyName || 'Story',
   Story,
@@ -30,4 +32,18 @@ describe('[Contextualbar Rendering]', () => {
       expect(results).toHaveNoViolations();
     },
   );
+
+  test('WithLongTitle should truncate long titles', () => {
+    const { getByRole } = render(<WithLongTitle />);
+    const titleElement = getByRole('heading');
+
+    const titleStyle = window.getComputedStyle(titleElement);
+    expect(titleStyle.overflow).toBe('hidden');
+    expect(titleStyle.textOverflow).toBe('ellipsis');
+    expect(titleStyle.whiteSpace).toBe('nowrap');
+
+    const flexContainer = titleElement.parentElement;
+    const containerStyle = window.getComputedStyle(flexContainer!);
+    expect(containerStyle.overflow).toBe('hidden');
+  });
 });
