+++ packages/fuselage/src/components/Contextualbar/ContextualbarAction.tsx
@@ -1,21 +1,13 @@
 import { memo } from 'react';
 
-import type { IconProps } from '..';
-import { IconButton } from '..';
-import type { IconButtonProps } from '../Button/IconButton';
+import { IconButton, type IconButtonProps } from '../Button';
 
-/**
- * @deprecated Use `ContextualbarV2ActionProps` instead.
- */
 export type ContextualbarActionProps = {
-  name: IconProps['name'];
-} & Omit<IconButtonProps, 'icon'>;
+  name: IconButtonProps['icon'];
+} & Omit<IconButtonProps, 'name' | 'icon'>;
 
 const ContextualbarAction = ({ name, ...props }: ContextualbarActionProps) => (
   <IconButton {...props} small flexShrink={0} icon={name} />
 );
 
-/**
- * @deprecated Use `ContextualbarV2Action` instead.
- */
 export default memo(ContextualbarAction);
