+++ packages/fuselage/src/components/Contextualbar/ContextualbarContent.tsx
@@ -3,9 +3,6 @@ import { memo } from 'react';
 
 import { Box, type BoxProps } from '../Box';
 
-/**
- * @deprecated Use `ContextualbarV2ContentProps` instead.
- */
 export type ContextualbarContentProps = Omit<BoxProps, 'ref'> &
   RefAttributes<HTMLElement>;
 
@@ -23,7 +20,4 @@ function ContextualbarContent(props: ContextualbarContentProps) {
   );
 }
 
-/**
- * @deprecated Use `ContextualbarV2Content` instead.
- */
 export default memo(ContextualbarContent);
