+++ packages/fuselage/src/components/Contextualbar/ContextualbarSection.tsx
@@ -3,9 +3,6 @@ import { memo } from 'react';
 
 import { Box, type BoxProps } from '../Box';
 
-/**
- * @deprecated Use `ContextualbarV2SectionProps` instead.
- */
 export type ContextualbarSectionProps = Omit<BoxProps, 'ref'> &
   RefAttributes<HTMLElement>;
 
@@ -13,8 +10,7 @@ function ContextualbarSection(props: ContextualbarSectionProps) {
   return (
     <Box
       rcx-vertical-bar__section
-      paddingInline={24}
-      paddingBlock={16}
+      padding={16}
       display='flex'
       alignItems='center'
       flexGrow={1}
@@ -25,7 +21,4 @@ function ContextualbarSection(props: ContextualbarSectionProps) {
   );
 }
 
-/**
- * @deprecated Use `ContextualbarV2Section` instead.
- */
 export default memo(ContextualbarSection);
