+++ packages/fuselage/src/components/Contextualbar/ContextualbarSkeleton.tsx
@@ -6,9 +6,6 @@ import { Skeleton } from '../Skeleton';
 import Contextualbar from './Contextualbar';
 import ContextualbarHeader from './ContextualbarHeader';
 
-/**
- * @deprecated Use `ContextualbarV2SkeletonProps` instead.
- */
 export type ContextualbarSkeletonProps = BoxProps;
 
 const ContextualbarSkeleton = (props: ContextualbarSkeletonProps) => (
@@ -27,7 +24,4 @@ const ContextualbarSkeleton = (props: ContextualbarSkeletonProps) => (
   </Contextualbar>
 );
 
-/**
- * @deprecated Use `ContextualbarV2Skeleton` instead.
- */
 export default memo(ContextualbarSkeleton);
