+++ packages/fuselage/src/components/Contextualbar/ContextualbarTitle.tsx
@@ -2,22 +2,17 @@ import { memo } from 'react';
 
 import { Box, type BoxProps } from '../Box';
 
-/**
- * @deprecated Use `ContextualbarV2TitleProps` instead.
- */
 export type ContextualbarTitleProps = BoxProps;
 
 const ContextualbarTitle = (props: ContextualbarTitleProps) => (
   <Box
     flexShrink={1}
     flexGrow={1}
-    fontScale='h4'
+    fontScale='h5'
     withTruncatedText
+    is='h5'
     {...props}
   />
 );
 
-/**
- * @deprecated Use `ContextualbarV2Title` instead.
- */
 export default memo(ContextualbarTitle);
