--- packages/fuselage/src/components/ContextualbarV2/ContextualbarButton.tsx
@@ -1,11 +1,0 @@
-import { memo } from 'react';
-
-import { Button, type ButtonProps } from '../Button';
-
-export type ContextualbarButtonProps = ButtonProps;
-
-const ContextualbarButton = (props: ContextualbarButtonProps) => (
-  <Button {...props} />
-);
-
-export default memo(ContextualbarButton);
