--- packages/fuselage/src/components/ContextualbarV2/index.ts
@@ -1,48 +1,0 @@
-export {
-  default as ContextualbarV2,
-  type ContextualbarProps as ContextualbarV2Props,
-} from './Contextualbar';
-export {
-  default as ContextualbarV2Action,
-  type ContextualbarActionProps as ContextualbarV2ActionProps,
-} from './ContextualbarAction';
-export {
-  default as ContextualbarV2Actions,
-  type ContextualbarActionsProps as ContextualbarV2ActionsProps,
-} from './ContextualbarActions';
-export {
-  default as ContextualbarV2Button,
-  type ContextualbarButtonProps as ContextualbarV2ButtonProps,
-} from './ContextualbarButton';
-export {
-  default as ContextualbarV2Content,
-  type ContextualbarContentProps as ContextualbarV2ContentProps,
-} from './ContextualbarContent';
-export {
-  default as ContextualbarV2EmptyContent,
-  type ContextualbarEmptyContentProps as ContextualbarV2EmptyContentProps,
-} from './ContextualbarEmptyContent';
-export {
-  default as ContextualbarV2Footer,
-  type ContextualbarFooterProps as ContextualbarV2FooterProps,
-} from './ContextualbarFooter';
-export {
-  default as ContextualbarV2Header,
-  type ContextualbarHeaderProps as ContextualbarV2HeaderProps,
-} from './ContextualbarHeader';
-export {
-  default as ContextualbarV2Icon,
-  type ContextualbarIconProps as ContextualbarV2IconProps,
-} from './ContextualbarIcon';
-export {
-  default as ContextualbarV2Section,
-  type ContextualbarSectionProps as ContextualbarV2SectionProps,
-} from './ContextualbarSection';
-export {
-  default as ContextualbarV2Skeleton,
-  type ContextualbarSkeletonProps as ContextualbarV2SkeletonProps,
-} from './ContextualbarSkeleton';
-export {
-  default as ContextualbarV2Title,
-  type ContextualbarTitleProps as ContextualbarV2TitleProps,
-} from './ContextualbarTitle';
