+++ packages/fuselage/src/components/index.ts
@@ -18,7 +18,6 @@ export * from './Chevron';
 export * from './Chip';
 export * from './CodeSnippet';
 export * from './Contextualbar';
-export * from './ContextualbarV2';
 export * from './Divider';
 export * from './Dropdown';
 export * from './EmailInput';
